Day 6 Part 2 in Prolog
In the end this was simple, having a triple nested list to represent all the answers confused me. The breakthrough was when I renamed it to be a "Group" rather than a "Person_List". Then it became evident that I keep the group constant as I iterated through the Questions. The main loop which handed groups off to be counted didn't need to change at all (I ended up renaming the vars to be Groups).